Added base Keil project
A simple Keil project is added as a starting point. The clock
configuration function in main is extracted from the Blinky example
code. Also the function SystemCoreClockUpdate is defined in
startup.c, which is a part of the Device Family Pack (DFP).
Thus, these functions will not be following the programming
conventions used in this project.